+++ migrations/20260131142526_create_cars.sql
@@ -1,11 +1,11 @@
 -- +goose Up
 -- +goose StatementBegin
 CREATE TABLE IF NOT EXISTS car (
-    car_id SERIAL PRIMARY KEY,
+    car_id BIGSERIAL PRIMARY KEY,
     mark VARCHAR(255),
     model VARCHAR(255),
-    owner_count INT,
-    price INT,
+    owner_count BIGINT,
+    price BIGINT,
     currency VARCHAR(3),
     options VARCHAR(255)[]
 );
